Fixed comparison compiler warning because of size_t compare. Removed performance issue with XML character encoding.

+++ agent/cppagent.cpp
@@ -209,7 +209,7 @@ int main(int aArgc, char *aArgv[])
       // Should have the format device:address:port
       string &adapter = *iter;
       string device, address, adapter_port;
-      int pos1 = adapter.find_first_of(':');
+      size_t pos1 = adapter.find_first_of(':');
       if (pos1 == string::npos) {
         cerr << "Bad format for adapter specification, must be: device:address:port" << endl;
         option_list.usage();
@@ -223,7 +223,7 @@ int main(int aArgc, char *aArgv[])
         option_list.usage();
       }
       
-      int pos2 = adapter.find_first_of(':', pos1 + 1);
+      size_t pos2 = adapter.find_first_of(':', pos1 + 1);
       if (pos2 == string::npos) {
         cerr << "Bad format for adapter specification, must be: device:address:port" << endl;
         option_list.usage();
